Module Translator_benchmarks.Unparsing_data

val name : string

Name of the benchmark

val info : string

Description of the benchmark

val tags : string list

Tags of the benchmark

type config

Configuration of the benchmark (eg sampling parameters, paths, etc)

val default_config : config

Default configuration of the benchmark

val config_encoding : config Tezos_base__TzPervasives.Data_encoding.t

Configuration encoding

type workload

Benchmark workload

val workload_encoding : workload Tezos_base__TzPervasives.Data_encoding.t

Workload encoding

val workload_to_vector : workload -> Tezos_benchmark.Sparse_vec.String.t

Optional conversion to vector, for report generation purposes

val models : (string * workload Tezos_benchmark.Model.t) list

Cost models

Benchmark generator

include Tezos_benchmark.Generator.S with type config := config and type workload := workload

val create_benchmarks : rng_state:Stdlib.Random.State.t -> bench_num:int -> config -> (unit -> workload Tezos_benchmark.Generator.benchmark) list

Creates a list of benchmarks, ready to be run. The bench_num option corresponds to the argument command-line specified by the user on command line, but it can be overriden by the config-specific settings. This is the case for instance when the benchmarks are performed on external artifacts. The benchmarks are thunked to prevent evaluating the workload until needed.